+++ GameServer.sln
@@ -5,8 +5,6 @@ VisualStudioVersion = 14.0.25123.0
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GameServer", "GameServer\GameServer.csproj", "{CB004023-B938-4EC6-B036-29EE2F55CFA4}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "InibinSharp", "InibinSharp\InibinSharp.csproj", "{F47AC3D4-6C0F-4F02-9A7F-FBED54CF6E0A}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ENetCS", "ENetCS\ENetCS.csproj", "{D3AFF7FA-96F6-42B8-B03C-E508DC80EC3A}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GameServerTests", "GameServerTests\GameServerTests.csproj", "{56E885A9-FC37-4ABB-B9CC-822DAC13CC67}"
@@ -27,14 +25,6 @@ Global
 		{CB004023-B938-4EC6-B036-29EE2F55CFA4}.Release|Any CPU.Build.0 = Release|Any CPU
 		{CB004023-B938-4EC6-B036-29EE2F55CFA4}.ReleaseKopiLua|Any CPU.ActiveCfg = Release|Any CPU
 		{CB004023-B938-4EC6-B036-29EE2F55CFA4}.ReleaseKopiLua|Any CPU.Build.0 = Release|Any CPU
-		{F47AC3D4-6C0F-4F02-9A7F-FBED54CF6E0A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F47AC3D4-6C0F-4F02-9A7F-FBED54CF6E0A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F47AC3D4-6C0F-4F02-9A7F-FBED54CF6E0A}.DebugKopiLua|Any CPU.ActiveCfg = Debug|Any CPU
-		{F47AC3D4-6C0F-4F02-9A7F-FBED54CF6E0A}.DebugKopiLua|Any CPU.Build.0 = Debug|Any CPU
-		{F47AC3D4-6C0F-4F02-9A7F-FBED54CF6E0A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F47AC3D4-6C0F-4F02-9A7F-FBED54CF6E0A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{F47AC3D4-6C0F-4F02-9A7F-FBED54CF6E0A}.ReleaseKopiLua|Any CPU.ActiveCfg = Release|Any CPU
-		{F47AC3D4-6C0F-4F02-9A7F-FBED54CF6E0A}.ReleaseKopiLua|Any CPU.Build.0 = Release|Any CPU
 		{D3AFF7FA-96F6-42B8-B03C-E508DC80EC3A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{D3AFF7FA-96F6-42B8-B03C-E508DC80EC3A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{D3AFF7FA-96F6-42B8-B03C-E508DC80EC3A}.DebugKopiLua|Any CPU.ActiveCfg = Debug|Any CPU
