--- InibinSharp/packages.config
@@ -1,8 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="System.Globalization" version="4.0.10" targetFramework="net46" />
-  <package id="System.Numerics.Vectors" version="4.1.0" targetFramework="net46" />
-  <package id="System.Resources.ResourceManager" version="4.0.0" targetFramework="net46" />
-  <package id="System.Runtime" version="4.0.20" targetFramework="net46" />
-  <package id="System.Runtime.Extensions" version="4.0.10" targetFramework="net46" />
-</packages>
